% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_tools.R
\name{calculate_10x_dbr}
\alias{calculate_10x_dbr}
\title{calculate_10x_dbr}
\usage{
calculate_10x_dbr(ncells)
}
\arguments{
\item{ncells}{number of cells captured}
}
\value{
a vector of doublet rates
}
\description{
calculate doublet rate for 10x sequencing runs
}
