% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_processing.R
\name{calculate_cellcycle}
\alias{calculate_cellcycle}
\title{calculate_cellcycle}
\usage{
calculate_cellcycle(x, org = "human", remove_genes = F, assay = "RNA")
}
\arguments{
\item{x}{Seurat object}

\item{org}{organism, either "human" or "mouse"}

\item{remove_genes}{If TRUE, remove genes from current assay after calculating cell cycle phase. Defaults to FALSE}

\item{assay}{assay name, defaults to "RNA"}
}
\description{
Calculate cell cycle score and predict cell cycle phase by Seurat
}
