% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_processing.R
\name{calculate_cluster_similarity}
\alias{calculate_cluster_similarity}
\title{calculate_cluster_similarity}
\usage{
calculate_cluster_similarity(
  x,
  cluster,
  assay = "RNA",
  slot = "data",
  variable.features = F
)
}
\arguments{
\item{x}{Seurat object}

\item{cluster}{cluster metadata column containing cluster labels}

\item{assay}{assay name for gene expression. Defaults to "RNA"}

\item{slot}{slot name for gene expression. Defaults to "data"}

\item{variable.features}{if TRUE, calculate correlation of variable features only}
}
\description{
calculate pearson correlation similarity between seurat clusters
}
