% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_processing.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{calculate_correlation}
\usage{
calculate_correlation(matrix)
}
\arguments{
\item{matrix}{gene expression matrix, clusters (columns) by features (rows)}
}
\description{
calculate pearson correlation of gene expression between columns of gene expression matrix
}
