% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_processing.R
\name{calculate_fractions}
\alias{calculate_fractions}
\title{calculate_fractions}
\usage{
calculate_fractions(x)
}
\arguments{
\item{x}{Seurat object}
}
\description{
Calculate fractions of mitochondrial, ribosomal, haemoglobin, TCR, BCR and MHC genes
}
