% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01-RNA_02-CELLCYCLE.R
\name{cc_by_seurat}
\alias{cc_by_seurat}
\title{Calculate cell cycle phase}
\usage{
cc_by_seurat(obj, split.by = NULL, assay = "RNA")
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{assay name, defaults to "RNA"}

\item{org}{Organism, either "human" or "mouse"}
}
\value{
Seurat object
}
\description{
This function calculates the cell cycle phase.
}
