% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_diffexp.R
\name{convert_deseq2_to_findmarkers}
\alias{convert_deseq2_to_findmarkers}
\title{convert_deseq2_to_findmarkers}
\usage{
convert_deseq2_to_findmarkers(res, x, assay = "RNA")
}
\description{
convert_deseq2_to_findmarkers
}
