% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_genesymbol.R
\name{convert_human_to_mouse}
\alias{convert_human_to_mouse}
\title{convert_human_to_mouse}
\usage{
convert_human_to_mouse(x, unique = T)
}
\arguments{
\item{x}{A vector of HGNC gene symbols.}

\item{unique}{Whether to disable one-to-many MGI gene symbols to be returned. Defaults to TRUE to disable one-to-many mapping (TRUE)}
}
\value{
If unique = TRUE, returns a vector of unique MGI gene symbols that are mapped one-to-one.
If unique = FALSE, returns a data frame with columns "MGI.symbol" and "HGNC.symbol" that are mapped one-to-many.
}
\description{
Convert human (HGNC) gene symbols to mouse (MGI) gene symbols; disable one-to-many mapping
}
