% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_genesymbol.R
\name{convert_mouse_to_human}
\alias{convert_mouse_to_human}
\title{convert_mouse_to_human}
\usage{
convert_mouse_to_human(x, unique = T)
}
\arguments{
\item{x}{A vector of MGI gene symbols.}

\item{unique}{Whether to disable one-to-many HGNC gene symbols to be returned. Defaults to TRUE to disable one-to-many mapping}
}
\value{
If unique = TRUE, returns a vector of unique HGNC gene symbols that are mapped one-to-one.
If unique = FALSE, returns a data frame with columns "MGI.symbol" and "HGNC.symbol" that are mapped one-to-many.
}
\description{
Convert mouse (MGI) gene symbols to human (HGNC) gene symbols; disable one-to-many mapping
}
