% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_genesymbol.R
\name{convert_mouse_to_human_seurat}
\alias{convert_mouse_to_human_seurat}
\title{convert_mouse_to_human_seurat}
\usage{
convert_mouse_to_human_seurat(x, assay = "RNA", unique = T)
}
\arguments{
\item{x}{Seurat object}

\item{assay}{assay to use}

\item{unique}{Whether to disable one-to-many HGNC gene symbols to be returned. Defaults to TRUE to disable one-to-many mapping}
}
\value{
If unique = TRUE, returns a vector of unique HGNC gene symbols that are mapped one-to-one.
If unique = FALSE, returns a data frame with columns "MGI.symbol" and "HGNC.symbol" that are mapped one-to-many.
}
\description{
Make a Seurat assay converting mouse (MGI) gene symbols to human (HGNC) gene symbols ; disable one-to-many mapping
}
