% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diffexp.R
\name{convert_seurat5_to_deseq2}
\alias{convert_seurat5_to_deseq2}
\title{Convert Seurat5 differential expression results to DESeq2 format}
\usage{
convert_seurat5_to_deseq2(diffexp)
}
\arguments{
\item{diffexp}{A dataframe with differential expression results}
}
\value{
A dataframe with differential expression results in DESeq2 format
}
\description{
This function converts differential expression results from Seurat5 to DESeq2 format.
}
\examples{
\dontrun{
convert_seurat5_to_deseq2(diffexp)
}
}
