% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_interoperability.R
\name{convert_seurat_to_sce}
\alias{convert_seurat_to_sce}
\title{convert_seurat_to_sce}
\usage{
convert_seurat_to_sce(x, assay = "RNA", return_genes = T)
}
\arguments{
\item{x}{Seurat object}

\item{assay}{assay name for gene expression. Defaults to "RNA"}

\item{return_genes}{if TRUE, return genes from assays named c("CC", "BCR", "TCR", "MHC") to current assay. Defaults to TRUE}
}
\description{
Convert Seurat to SingleCellExperiment object with SeuratWrapper::as.SingleCellExperiment()
}
