% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_objects.R
\name{create_seurat_object}
\alias{create_seurat_object}
\title{create_seurat_object}
\usage{
create_seurat_object(
  dir,
  project,
  file_format = "cellranger",
  hto_prefix = NULL,
  merge = TRUE
)
}
\arguments{
\item{dir}{A vector of directories containing cellranger/h5 outputs}

\item{project}{A vector of run IDs for each directory}

\item{file_format}{The format of the data to be loaded. Either "cellranger" or "h5". Defaults to "cellranger"}

\item{hto_prefix}{Prefix to identify HTO tag names if HTO library is present. Defaults to NULL}
}
\value{
A list of Seurat objects, with each entry corresponding to a sample
}
\description{
Creates a list of Seurat objects from 10x Genomics cellranger/h5 outputs.
This function processes RNA, ADT (Antibody-Derived Tags), and HTO (Hashtag Oligonucleotides)
modalities if present in the cellranger/h5 outputs.
}
