% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{dotplot_add_statistics}
\alias{dotplot_add_statistics}
\title{Add differential expression significance to expression data}
\usage{
dotplot_add_statistics(
  obj,
  exprs,
  diffexp,
  fc.thresh = 0,
  padj.thresh = 0.05,
  min.pct = 0.1,
  split.by = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{exprs}{Expression dataframe}

\item{diffexp}{Differential expression results}

\item{split.by}{Column to split by (optional)}
}
\value{
Annotated expression dataframe
}
\description{
Adds statistical significance information from differential expression results
}
\keyword{internal}
