% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{dotplot_convert_genes}
\alias{dotplot_convert_genes}
\title{Format feature list for dotplot}
\usage{
dotplot_convert_genes(input, unique = FALSE)
}
\arguments{
\item{input}{Genes as character vector or named list}

\item{unique}{Logical; whether to keep only unique genes across gene sets}
}
\value{
Dataframe with formatted features
}
\description{
Processes a list or vector of gene names into a formatted dataframe for dot plots
}
\keyword{internal}
