% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{dotplot_format_features}
\alias{dotplot_format_features}
\title{Format gene list for dotplot}
\usage{
dotplot_format_features(input, unique = FALSE, coord_flip = FALSE)
}
\arguments{
\item{input}{Features as character vector or named list}

\item{unique}{Logical; whether to keep only unique features across gene sets}
}
\value{
Dataframe with formatted features
}
\description{
Processes a list or vector of gene names into a formatted dataframe for dot plots
}
\keyword{internal}
