% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{dotplot_get_expression}
\alias{dotplot_get_expression}
\title{Get gene expression data}
\usage{
dotplot_get_expression(
  obj,
  features,
  group.by,
  split.by = NULL,
  scale = TRUE,
  mode = 1
)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{Vector of features to include}

\item{group.by}{Column to group by}

\item{split.by}{Column to split by (optional)}

\item{scale}{Logical; whether to scale data}

\item{mode}{1 or 2; 1 is to replace them with NA, 2 is to replace them with 0}
}
\value{
Dataframe with summarized expression values
}
\description{
Extracts and summarizes gene expression data from a Seurat object
}
\keyword{internal}
