% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{dotplot_validate_features}
\alias{dotplot_validate_features}
\title{Validate features}
\usage{
dotplot_validate_features(obj, input_df)
}
\arguments{
\item{obj}{Seurat object}

\item{input_df}{Dataframe with formatted features}
}
\value{
Dataframe with validated features
}
\description{
Validate features for dotplot
}
\keyword{internal}
