% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01-RNA_01-QC.R
\name{feature_percentages}
\alias{feature_percentages}
\title{feature percentages}
\usage{
feature_percentages(obj, assay = "RNA")
}
\arguments{
\item{obj}{Seurat object}
}
\value{
Seurat object
}
\description{
This function calculates the percentage of features in the Seurat object.
}
