% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c-DIMREDUC_utils.R
\name{find_npcs}
\alias{find_npcs}
\title{find_npcs}
\usage{
find_npcs(obj, reduction = "pca")
}
\arguments{
\item{obj}{Seurat object}

\item{reduction}{Reduction name}
}
\value{
Number of PCs
}
\description{
find_npcs
}
