% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_assays.R
\name{get_assay_keys}
\alias{get_assay_keys}
\title{get_assay_keys}
\usage{
get_assay_keys(obj)
}
\arguments{
\item{obj}{A Seurat object}
}
\value{
A named vector of assay keys
}
\description{
Retrieves all assay keys from a Seurat object.
This function extracts the key attribute from each assay in the Seurat object.
}
\keyword{internal}
