% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_genesymbol.R
\name{get_biomart}
\alias{get_biomart}
\title{get_biomart}
\usage{
get_biomart(host = "https://dec2021.archive.ensembl.org")
}
\arguments{
\item{host}{URL to retrieve archived ensembl biomarts}
}
\value{
set human and mouse biomart objects in local environment
}
\description{
Initializes the Ensembl Mart for human and mouse gene symbols.
}
