% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduction.R
\name{get_embeddings}
\alias{get_embeddings}
\title{get_embeddings}
\usage{
get_embeddings(obj, reduction)
}
\arguments{
\item{obj}{Seurat object}

\item{reduction}{Reduction name}
}
\value{
Data frame with embeddings
}
\description{
Get embeddings from Seurat object
}
\keyword{internal}
