% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_functions.R
\name{get_name}
\alias{get_name}
\title{Get the name of an object}
\usage{
get_name(obj)
}
\arguments{
\item{obj}{The object to search for}
}
\value{
The name of the object, or NULL if not found
}
\description{
This function searches the global environment for an object and returns its name.
}
\examples{
\dontrun{
get_name(my_plot)
}
}
