% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_assays.R
\name{get_org}
\alias{get_org}
\title{get_org}
\usage{
get_org(obj, assay = "RNA")
}
\arguments{
\item{obj}{A Seurat object}
}
\value{
The organism
}
\description{
Retrieves the organism from a Seurat object.
}
