% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting_reduction.R
\name{gg_colorlabel}
\alias{gg_colorlabel}
\title{gg_colorlabel}
\usage{
gg_colorlabel(plot, df, group.by, label.size = 4)
}
\arguments{
\item{plot}{ggplot object}

\item{df}{Data frame}

\item{group.by}{Metadata column to group by}

\item{label.size}{Label size}

\item{label}{Metadata column to label}

\item{label.color}{Label color}
}
\value{
ggplot object
}
\description{
Add labels to ggplot
}
