% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduction.R
\name{gg_count}
\alias{gg_count}
\title{gg_count}
\usage{
gg_count(plot, df, split.by = NULL, count.text.size = 5)
}
\arguments{
\item{plot}{ggplot object}

\item{df}{Data frame}

\item{split.by}{Metadata column to split by}

\item{count.text.size}{Text size for cell count}
}
\value{
ggplot object
}
\description{
Add total cell count to ggplot
}
\keyword{internal}
