% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{gg_dotplot}
\alias{gg_dotplot}
\title{Create a dot plot with customizable settings}
\usage{
gg_dotplot(exprs, x, y, coord_flip = TRUE, palette, scale = TRUE)
}
\arguments{
\item{exprs}{Dataframe with expression values}

\item{x}{X-axis column name}

\item{y}{Y-axis column name}

\item{coord_flip}{Logical; whether the plot coordinates are flipped}

\item{palette}{Color palette for expression values}

\item{scale}{Logical; whether data was scaled}
}
\value{
A ggplot object
}
\description{
Create a dot plot with customizable settings
}
\keyword{internal}
