% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting_reduction.R
\name{gg_mask}
\alias{gg_mask}
\title{gg_mask}
\usage{
gg_mask(plot, df, group.by, mask.size = 1)
}
\arguments{
\item{plot}{ggplot object}

\item{df}{Data frame}

\item{group.by}{Metadata column to group by}

\item{mask.size}{Mask size}
}
\value{
ggplot object
}
\description{
Add mask to ggplot
}
