% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting_reduction.R
\name{gg_point}
\alias{gg_point}
\title{gg_point}
\usage{
gg_point(
  plot,
  df,
  group.by = NULL,
  pt.size = 0.8,
  pt.alpha = 0.5,
  legend.ncol = 1
)
}
\arguments{
\item{plot}{ggplot object}

\item{df}{Data frame}

\item{group.by}{Metadata column to group by}

\item{pt.size}{Point size}

\item{pt.alpha}{Point alpha}

\item{legend.ncol}{Number of columns for legend}
}
\value{
ggplot object
}
\description{
Add points to ggplot
}
