% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting_reduction.R
\name{gg_reduction}
\alias{gg_reduction}
\title{gg_reduction}
\usage{
gg_reduction(
  df,
  reduction.type = NULL,
  group.by = NULL,
  split.by = NULL,
  facet.ncol = 2,
  shuffle = T
)
}
\arguments{
\item{df}{Data frame}

\item{reduction.type}{Reduction type}

\item{group.by}{Metadata column to group by}

\item{split.by}{Metadata column to split by}

\item{facet.ncol}{Number of columns for facets}

\item{shuffle}{Whether to shuffle points}
}
\value{
Data frame with reduction
}
\description{
Get reduction from Seurat object
}
