% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c-DIMREDUC_utils.R
\name{hvf_by_seurat_vst}
\alias{hvf_by_seurat_vst}
\title{hvf_by_seurat_vst}
\usage{
hvf_by_seurat_vst(
  obj,
  normalization.method = "SCT",
  split.by = NULL,
  nfeatures = 2000,
  output_dir = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{normalization.method}{Normalization method}

\item{split.by}{Split by}

\item{nfeatures}{Number of features}

\item{output_dir}{Output directory}
}
\value{
Seurat object
}
\description{
hvf_by_seurat_vst
}
