% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01d-INTEGRATION_utils.R
\name{int_by_seurat}
\alias{int_by_seurat}
\title{integrate_by_seurat}
\usage{
int_by_seurat(
  obj,
  split.by = NULL,
  normalization.method = "SCT",
  vars.to.regress = NULL,
  method = "rpca",
  return.only.var.genes = T,
  k.weight = 100
)
}
\arguments{
\item{obj}{Seurat object}

\item{split.by}{Split by}

\item{normalization.method}{Normalization method}

\item{vars.to.regress}{Variables to regress out}

\item{method}{Method for integration}

\item{return.only.var.genes}{Return only variable genes}

\item{k.weight}{K weight}
}
\value{
Seurat object
}
\description{
integrate_by_seurat
}
