% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_processing.R
\name{integrate_v4}
\alias{integrate_v4}
\title{integrate_v4}
\usage{
integrate_v4(
  x,
  split.by,
  assay = "RNA",
  nfeatures = 3000,
  npcs = 50,
  vars.to.regress = NULL,
  method = "rpca",
  k.weight = 100
)
}
\arguments{
\item{x}{Seurat object}

\item{split.by}{variable to split object by}

\item{assay}{assay name of gene expression. Defaults to "RNA"}

\item{nfeatures}{no. of features to select for integration}

\item{npcs}{no. of PCs to computec}

\item{vars.to.regress}{variables to regress}

\item{method}{integration method, see Seurat::indIntegrationAnchors()}

\item{k.weight}{k.weight, see Seurat::IntegrateData(), reduce when no. of cells in samples are low.}
}
\description{
Wrapper function to integrate cells by samples
}
