% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_features.R
\name{intersect_features}
\alias{intersect_features}
\title{intersect_features}
\usage{
intersect_features(obj, split.by, assay)
}
\arguments{
\item{assay}{assay name}

\item{obj.list}{a list of Seurat object}

\item{min.cells}{minimum number of cells expressing the gene}
}
\description{
intersect common features between a list of seurat objects
}
