% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_tools.R
\name{intersect_genes}
\alias{intersect_genes}
\title{intersect_genes}
\usage{
intersect_genes(x, assay = "RNA", min.cells = 3)
}
\arguments{
\item{x}{a list of Seurat object}

\item{assay}{assay name}

\item{min.cells}{minimum number of cells expressing the gene}
}
\description{
intersect common gene names between a list of seurat objects
}
