% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_assays.R
\name{join_layers}
\alias{join_layers}
\title{join_layers}
\usage{
join_layers(obj)
}
\arguments{
\item{obj}{Seurat object}
}
\value{
Seurat object
}
\description{
Join layers for all Assay5 objects in a Seurat object
}
