% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_tools.R
\name{list_to_df}
\alias{list_to_df}
\title{list_to_df}
\usage{
list_to_df(list)
}
\arguments{
\item{list}{A list of vectors}
}
\value{
A dataframe where each column is a vector from the list
}
\description{
Convert list to dataframe
}
