% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_logging.R
\name{log_progress_bar}
\alias{log_progress_bar}
\title{Log progress bar}
\usage{
log_progress_bar(i, n, pb = NULL)
}
\arguments{
\item{i}{The current iteration}

\item{n}{The total number of iterations}

\item{pb}{The progress bar}
}
\value{
The progress bar
}
\description{
Log progress bar
}
