% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_vdj.R
\name{match_cdr3}
\alias{match_cdr3}
\title{match_cdr3}
\usage{
match_cdr3(
  query,
  reference,
  method = "hamming",
  match = c("v_call_VDJ_main", "j_call_VDJ_main"),
  ncores = 1
)
}
\arguments{
\item{query}{query dataframe with "junction_aa_VDJ" as column}

\item{reference}{reference dataframe with "junction_aa_VDJ" as column}

\item{method}{method to calculate CDR3 AA distance : "hamming" or "levenshtein"}

\item{match}{additional metadata columns to match}

\item{ncores}{no. of cores}
}
\description{
match CDR3 amino acid sequence
}
