% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_assays.R
\name{merge_assays}
\alias{merge_assays}
\title{merge_assays}
\usage{
merge_assays(obj, assay1, assay2)
}
\arguments{
\item{obj}{Seurat object containing both assays}

\item{assay1}{Name of the assay that will contain the merged data}

\item{assay2}{Name of the assay that will be merged into assay1 and then removed}
}
\value{
Seurat object with merged assays
}
\description{
Merge two Assay5 objects in a Seurat object, combining their features
}
