% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_objects.R
\name{merge_objects}
\alias{merge_objects}
\title{merge_objects}
\usage{
merge_objects(obj.list)
}
\arguments{
\item{obj.list}{A list of Seurat objects.}
}
\value{
A single Seurat object.
}
\description{
Merges a list of Seurat objects into a single Seurat object.
}
