% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_tools.R
\name{metadata_checkpoint}
\alias{metadata_checkpoint}
\title{metadata_checkpoint}
\usage{
metadata_checkpoint(x, dir = "seurat/0_metadata/", save = F, load = NULL)
}
\arguments{
\item{x}{Seurat Object}

\item{dir}{directory storing metadata}

\item{save}{save metadata as tsv}

\item{load}{specific metadata.tsv to load}
}
\value{
Seurat Object with updated metadata file
}
\description{
add checkpoint for metadata
}
