% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot boxplot}
\usage{
plot_boxplot(
  df,
  var,
  group.by,
  split.by = NULL,
  filename = NULL,
  output_dir = NULL
)
}
\arguments{
\item{df}{Data frame}

\item{var}{Metadata column to plot}

\item{group.by}{Metadata column to group by}

\item{split.by}{Metadata column to split the object by}

\item{filename}{Name of the file to save}

\item{output_dir}{Directory to save the output}
}
\value{
ggplot object
}
\description{
Plot boxplot
}
