% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b-CELLCYCLE_plot.R
\name{plot_cc}
\alias{plot_cc}
\title{Plot cell cycle}
\usage{
plot_cc(obj, var = "Phase", split.by = NULL, output_dir = NULL)
}
\arguments{
\item{obj}{Seurat object}

\item{split.by}{Metadata column to split the object by}

\item{output_dir}{Directory to save the output}
}
\value{
List of plots
}
\description{
Plot cell cycle
}
