% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_vdj.R
\name{plot_clone_number}
\alias{plot_clone_number}
\title{plot_clone_number}
\usage{
plot_clone_number(x, group.by, cols = NULL)
}
\arguments{
\item{x}{Seurat object}

\item{group.by}{column to group cells by}

\item{cols}{colors}
}
\description{
plot clone number
}
