% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plotting.R
\name{plot_density_umap}
\alias{plot_density_umap}
\title{plot_density_umap}
\usage{
plot_density_umap(
  x,
  reduction = "umap",
  split.by,
  adjust = 1,
  contour = T,
  together = F,
  count = T,
  pt.size = 0.5
)
}
\arguments{
\item{x}{Seurat object}

\item{reduction}{reduction name, defaults to "umap"}

\item{split.by}{metadata column to group by}

\item{adjust}{adjust for stat_density_2d() or geom_density_2d}

\item{contour}{plot bins for geom_density_2d if TRUE}

\item{together}{plot cells from all facet in background if TRUE}

\item{count}{display the number of cells if TRUE}

\item{pt.size}{geom_point size}
}
\description{
Plot DensityUMAP from Seurat object
}
