% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diffexp.R
\name{plot_diffexp_comparison}
\alias{plot_diffexp_comparison}
\title{Plot differential expression comparison}
\usage{
plot_diffexp_comparison(
  df,
  x,
  y,
  thresh = 0.5,
  highlight.genes = c(),
  tr = "Both Up",
  bl = "Both Down",
  tl = "Opposite",
  br = "Opposite"
)
}
\arguments{
\item{df}{A dataframe with differential expression results}

\item{x}{The column name of the first condition}

\item{y}{The column name of the second condition}

\item{thresh}{A numeric threshold for the padj column}

\item{highlight.genes}{A vector of gene IDs to highlight in the plot}

\item{tr}{The label for the top right quadrant}

\item{bl}{The label for the bottom left quadrant}

\item{tl}{The label for the top left quadrant}

\item{br}{The label for the bottom right quadrant}
}
\value{
A ggplot object
}
\description{
This function plots differential expression results between two conditions
and highlights significant genes in a scatter plot.
}
\examples{
\dontrun{
plot_diffexp_comparison(df, x = "log2FoldChange_condition1", y = "log2FoldChange_condition2", thresh = 0.5, highlight.genes = c("gene1", "gene2"), tr = "Both Up", bl = "Both Down", tl = "Opposite", br = "Opposite")
}

}
