% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting_dotplot.R
\name{plot_dotplot}
\alias{plot_dotplot}
\title{plot_dotplot}
\usage{
plot_dotplot(
  object,
  features,
  assay = NULL,
  cols = get_palette("Reds", n = 7)[c(1, 7)],
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  base_size = 16,
  dot.scale = 14,
  idents = NULL,
  group.by = NULL,
  split.by = NULL,
  cluster.idents = FALSE,
  scale = TRUE,
  scale.by = "radius",
  scale.min = NA,
  scale.max = NA,
  plot.title = " ",
  diffexp = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Features to plot}

\item{assay}{Assay to use}

\item{cols}{Colors to use}

\item{col.min}{Minimum value for color scale}

\item{col.max}{Maximum value for color scale}

\item{dot.min}{Minimum value for dot size}

\item{dot.scale}{Dot scale}

\item{idents}{Identities to plot}

\item{group.by}{Group by}

\item{split.by}{Split by}

\item{cluster.idents}{Cluster identities}

\item{scale}{Scale}

\item{scale.by}{Scale by}

\item{scale.min}{Minimum value for scale}

\item{scale.max}{Maximum value for scale}

\item{plot.title}{Plot title}
}
\value{
A ggplot object
}
\description{
plot_dotplot
}
