% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plotting.R
\name{plot_feature}
\alias{plot_feature}
\title{plot_feature}
\usage{
plot_feature(
  x,
  assay = "RNA",
  features,
  palette = "viridis",
  direction = 1,
  ncol,
  plot = F,
  ...
)
}
\arguments{
\item{x}{Seurat object}

\item{features}{feature names stored in a vector}

\item{palette}{a vector of hexcode colors, or palettes from "RColorBrewer" or "viridis"}

\item{direction}{direction of the palette vector}

\item{ncol}{no. of columns for the plots}

\item{plot}{If TRUE, plot with cowplot::plot_grid(). Defaults to FALSE to return a list of ggplot objects}

\item{...}{arguments for Seurat::FeaturePlot()}
}
\description{
Plot FeaturePlot from Seurat object
}
