% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a-QC_plot.R
\name{plot_feature_percentages}
\alias{plot_feature_percentages}
\title{Plot feature percentages}
\usage{
plot_feature_percentages(
  obj,
  assay = "RNA",
  threshold = NULL,
  split.by = NULL,
  output_dir = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{Assay name}

\item{threshold}{Threshold for each feature}

\item{split.by}{Metadata column to split the object by}

\item{output_dir}{Directory to save the output}
}
\value{
List of plots
}
\description{
Plot feature percentages
}
