% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c-DIMREDUC_plot.R
\name{plot_featureloadings}
\alias{plot_featureloadings}
\title{plot_featureloadings}
\usage{
plot_featureloadings(obj, reduction = "pca", output_dir = NULL)
}
\arguments{
\item{obj}{seurat object}

\item{reduction}{reduction name}

\item{output_dir}{output directory}
}
\value{
ggplot
}
\description{
plot_featureloadings
}
