% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plotting.R
\name{plot_hvf}
\alias{plot_hvf}
\title{plot_hvf}
\usage{
plot_hvf(x, assay = "RNA", nfeatures = 3000, n = 10)
}
\arguments{
\item{x}{Seurat object}

\item{assay}{assay name}

\item{nfeatures}{no. of features to select}

\item{n}{number of features to label}
}
\description{
Plot highly variable features from Seurat object
}
