% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting.R
\name{plot_percent}
\alias{plot_percent}
\title{Plot percent}
\usage{
plot_percent(
  df,
  group.by,
  var,
  split.by = NULL,
  pal = NULL,
  legend.ncol = 1,
  count = T,
  filename = NULL,
  output_dir = NULL
)
}
\arguments{
\item{df}{Data frame}

\item{group.by}{Metadata column to group by}

\item{var}{Metadata column to plot}

\item{split.by}{Metadata column to split the object by}

\item{pal}{Palette to use}

\item{legend.ncol}{Number of columns in the legend}

\item{count}{Whether to show the count of each group}

\item{filename}{Name of the file to save}

\item{output_dir}{Directory to save the output}
}
\value{
ggplot object
}
\description{
Plot percent
}
