% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a-QC_plot.R
\name{plot_qc_cell_count}
\alias{plot_qc_cell_count}
\title{Plot cell count}
\usage{
plot_qc_cell_count(obj, split.by = NULL, output_dir = NULL)
}
\arguments{
\item{obj}{Seurat object}

\item{split.by}{Metadata column to split the object by}

\item{output_dir}{Directory to save the output}
}
\value{
ggplot object
}
\description{
Plot cell count
}
