% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting_reduction.R
\name{plot_reduction}
\alias{plot_reduction}
\title{plot_reduction}
\usage{
plot_reduction(
  obj,
  reduction = "umap",
  reduction.type = "UMAP",
  group.by,
  split.by = NULL,
  facet.ncol = 2,
  shuffle = T,
  ...
)
}
\arguments{
\item{obj}{Seurat object}

\item{reduction}{Reduction name}

\item{reduction.type}{Reduction type}

\item{group.by}{Metadata column to group by}

\item{split.by}{Metadata column to split by}

\item{facet.ncol}{Number of columns for facets}

\item{shuffle}{Whether to shuffle points}

\item{...}{Additional arguments to gg_features}
}
\value{
ggplot object
}
\description{
Plot reduction from Seurat object
}
\examples{
\dontrun{
plot_reduction(obj, group.by = "seurat_clusters")
}
}
