% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting_reduction.R
\name{plot_reduction_mapquery}
\alias{plot_reduction_mapquery}
\title{plot_reduction_mapquery()}
\usage{
plot_reduction_mapquery(
  obj,
  ref.obj,
  group.by = NULL,
  split.by = NULL,
  facet.ncol = 2,
  reduction = "umap",
  reduction.type = "UMAP",
  shuffle = T,
  ...
)
}
\arguments{
\item{obj}{Seurat object}

\item{group.by}{Metadata column to group by}

\item{facet.ncol}{Number of columns for facets}

\item{reduction}{Reduction name}

\item{shuffle}{Whether to shuffle points}

\item{...}{Additional arguments to plot_reduction}
}
\value{
ggplot object
}
\description{
Plot reduction from Seurat object with mapquery
}
\examples{
\dontrun{
plot_reduction_mapquery(obj, group.by = "seurat_clusters")
}
}
