% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduction.R
\name{plot_reduction_mask}
\alias{plot_reduction_mask}
\title{plot_reduction_mask}
\usage{
plot_reduction_mask(obj, group.by, reduction = "umap", mask.size = 1, ...)
}
\arguments{
\item{obj}{Seurat object}

\item{group.by}{Metadata column to group by}

\item{reduction}{Reduction name}

\item{mask.size}{Mask size}

\item{...}{Additional arguments to plot_reduction}
}
\value{
ggplot object
}
\description{
Plot reduction from Seurat object with masks
}
\examples{
\dontrun{
plot_reduction_mask(obj, group.by = "seurat_clusters")
}
}
