% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plotting.R
\name{plot_sc_quality}
\alias{plot_sc_quality}
\title{plot_sc_quality}
\usage{
plot_sc_quality(x, group.by, assay = "RNA", cols = NULL)
}
\arguments{
\item{x}{Seurat object}

\item{group.by}{column to group by}

\item{assay}{assay to use}

\item{cols}{colors}
}
\description{
quality control plot for single cell
}
