% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_vdj.R
\name{plot_shm}
\alias{plot_shm}
\title{plot_shm}
\usage{
plot_shm(x, group.by, facet.by = NULL, cols = NULL, adjust = 1)
}
\arguments{
\item{x}{Seurat object}

\item{group.by}{column to group cells by}

\item{facet.by}{variable to facet by}

\item{cols}{colors}

\item{adjust}{adjust}
}
\description{
plot somatic hypermuation
}
