% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plotting.R
\name{plot_similarity_heatmap}
\alias{plot_similarity_heatmap}
\title{plot_similarity_heatmap}
\usage{
plot_similarity_heatmap(
  correlation.matrix,
  annotations = colnames(correlation.matrix),
  color = NULL,
  cluster = T
)
}
\arguments{
\item{correlation.matrix}{a matrix of pearson correlation score}

\item{annotations}{a vector of cluster annotations for ComplexHeatmap::HeatmapAnnotation(). Defaults to colnames(correlation.matrix)}

\item{color}{colors for cluster annotations}
}
\description{
plot similarity heatmap between seurat clusters
}
