% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_vdj.R
\name{plot_vdj}
\alias{plot_vdj}
\title{plot_vdj}
\usage{
plot_vdj(x, variable = "isotype", ...)
}
\arguments{
\item{x}{Seurat object}

\item{variable}{metadata column to plot : "vdj_qc", "isotype", "clone_by_count", "clone_by_percent"}

\item{...}{refer to plot_percent}
}
\description{
plot vdj information for cells
}
