% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_02-HTO.R
\name{process_hto}
\alias{process_hto}
\title{process_hto}
\usage{
process_hto(obj, assay = "HTO")
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{name of HTO assay, defaults to "HTO"}
}
\description{
Run clustering algorithms on HTO counts
}
