% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-RNA.R
\name{process_RNA_CELLCYCLE}
\alias{process_RNA_CELLCYCLE}
\title{Process cell cycle scoring}
\usage{
process_RNA_CELLCYCLE(obj, assay = "RNA", split.by = NULL, output_dir = NULL)
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{Assay to validate}

\item{split.by}{Metadata column to split the object by}

\item{output_dir}{Directory to save the output}
}
\value{
Seurat object
}
\description{
Process cell cycle scoring
}
