% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-RNA.R
\name{process_RNA_DIMREDUC}
\alias{process_RNA_DIMREDUC}
\title{Process RNA dimensionality reduction}
\usage{
process_RNA_DIMREDUC(obj, assay = "RNA", split.by = NULL, output_dir = NULL)
}
\description{
Process RNA dimensionality reduction
}
