% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_processing.R
\name{process_seurat}
\alias{process_seurat}
\title{process_seurat}
\usage{
process_seurat(
  x,
  assay = "RNA",
  nfeatures = 2000,
  dims = 1:10,
  res = 0.4,
  samples = NULL,
  vars.to.regress,
  ...
)
}
\arguments{
\item{x}{Seurat object}

\item{assay}{assay name of gene expression. Defaults to "RNA"}

\item{nfeatures}{number of variable features to use. Defaults to 2000}

\item{dims}{no. of PCs to use. Defaults to 1:10}

\item{res}{clustering resolution. Defaults to 0.4}

\item{samples}{sample column}

\item{vars.to.regress}{variables to regress}
}
\description{
Wrapper function to cluster cells by gene expression
}
