% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a-QC_util.R
\name{qc_by_mad}
\alias{qc_by_mad}
\title{qc by MAD}
\usage{
qc_by_mad(
  obj,
  assay = "RNA",
  split.by = NULL,
  var = c("percent.mt", "percent.hb"),
  dev = 5,
  output_dir = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{Assay to validate}

\item{split.by}{Metadata column to split the object by}

\item{var}{Variable columns from meta.data to calculate MAD for}

\item{dev}{Deviation from median}

\item{output_dir}{Directory to save MAD threshold table and outcomes}
}
\value{
Seurat object
}
\description{
This function calculates the MAD of features in the Seurat object.
}
