% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a-QC_util.R
\name{qc_by_standard}
\alias{qc_by_standard}
\title{qc standard}
\usage{
qc_by_standard(
  obj,
  split.by = NULL,
  assay = "RNA",
  min.features = 200,
  min.pct = 0.01,
  min.cells = NULL,
  output_dir = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{split.by}{Metadata column to split the object by}

\item{assay}{Assay to validate}

\item{min.features}{Minimum number of features to keep}

\item{min.pct}{Minimum percentage of cells expressing a feature}

\item{min.cells}{Minimum number of cells expressing a feature}
}
\value{
Seurat object
}
\description{
This function performs standard quality control on a Seurat object.
}
