% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a-QC_util.R
\name{qc_find_percentages}
\alias{qc_find_percentages}
\title{feature percentages}
\usage{
qc_find_percentages(obj, assay = "RNA")
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{Assay to validate}
}
\value{
Seurat object
}
\description{
This function calculates the percentage of features in the Seurat object.
}
