% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_features.R
\name{remove_features}
\alias{remove_features}
\title{remove_features}
\usage{
remove_features(obj, assay, features)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{a vector of features to remove}
}
\description{
remove features from a Seurat object
}
