% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_features.R
\name{remove_vdj_features}
\alias{remove_vdj_features}
\title{remove_vdj_features}
\usage{
remove_vdj_features(obj, bcr = T, tcr = T, from.assay = "RNA")
}
\arguments{
\item{obj}{Seurat object}

\item{bcr}{if TRUE, remove BCR-VDJ features. Defaults to FALSE}

\item{tcr}{if TRUE, remove TCR-VDJ features. Defaults to FALSE}

\item{from.assay}{current assay name, defaults to "RNA"}
}
\description{
Remove VDJ features from an assay
}
