% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_tools.R
\name{remove_vdj_genes}
\alias{remove_vdj_genes}
\title{remove_vdj_genes}
\usage{
remove_vdj_genes(x, bcr = T, tcr = T, from.assay = "RNA")
}
\arguments{
\item{x}{Seurat object}

\item{bcr}{if TRUE, remove BCR-VDJ genes. Defaults to FALSE}

\item{tcr}{if TRUE, remove TCR-VDJ genes. Defaults to FALSE}

\item{from.assay}{current assay name, defaults to "RNA"}
}
\description{
Remove VDJ genes from an assay
}
