% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_features.R
\name{return_features}
\alias{return_features}
\title{return_features}
\usage{
return_features(obj, from.assay, to.assay = "RNA")
}
\arguments{
\item{obj}{Seurat object}

\item{from.assay}{assay name from which features are taken}

\item{to.assay}{assay name to which features are stored, defaults to "RNA"}
}
\description{
Return features from an assay to a specific assay
}
