% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_tools.R
\name{return_genes}
\alias{return_genes}
\title{return_genes}
\usage{
return_genes(x, from.assay, to.assay = "RNA")
}
\arguments{
\item{x}{Seurat object}

\item{from.assay}{assay name from which genes/features are taken}

\item{to.assay}{assay name to which genes/features are stored, defaults to "RNA"}
}
\description{
Return genes from an assay to a specific assay
}
