% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_02-HTO.R
\name{run_htodemux}
\alias{run_htodemux}
\title{run_htodemux}
\usage{
run_htodemux(obj, assay = "HTO", save_dir)
}
\arguments{
\item{obj}{Seurat object}

\item{process}{if TRUE, run cluster algorithms on HTO counts}
}
\description{
HTODemux algorithm to demultiplex cell hashtags
}
