% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_02-HTO.R
\name{run_multiseqdemux}
\alias{run_multiseqdemux}
\title{run_multiseqdemux}
\usage{
run_multiseqdemux(obj, process = F)
}
\arguments{
\item{obj}{Seurat object}

\item{process}{if TRUE, run cluster algorithms on HTO counts}
}
\description{
MULTIseqDemux algorithm to demultiplex cell hashtags
}
