% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c-DIMREDUC_utils.R
\name{run_pca}
\alias{run_pca}
\title{run_pca}
\usage{
run_pca(obj, reduction = "pca", output_dir = NULL)
}
\arguments{
\item{obj}{Seurat object}

\item{reduction}{Reduction name}

\item{output_dir}{Output directory}
}
\value{
Seurat object
}
\description{
run_pca
}
