% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c-DIMREDUC_utils.R
\name{run_umap}
\alias{run_umap}
\title{run_umap}
\usage{
run_umap(
  obj,
  split.by = NULL,
  reduction = "pca",
  umap.name = "umap",
  npcs = NULL,
  output_dir = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{split.by}{Split by}

\item{reduction}{Reduction name}

\item{output_dir}{Output directory}
}
\value{
Seurat object
}
\description{
run_umap
}
