% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_functions.R
\name{save_jupyter_plot}
\alias{save_jupyter_plot}
\title{Save a plot in Jupyter notebook}
\usage{
save_jupyter_plot(
  plot,
  save_dir,
  width = getOption("repr.plot.width"),
  height = getOption("repr.plot.height")
)
}
\arguments{
\item{plot}{The plot to save}

\item{save_dir}{The directory to save the plot}

\item{width}{The width of the plot}

\item{height}{The height of the plot}
}
\value{
None. Creates a PDF file in the specified directory
}
\description{
This function saves a plot as a PDF file in the specified directory.
It uses the Cairo PDF device if available, or falls back to standard PDF if not.
}
\examples{
\dontrun{
save_jupyter_plot(my_plot, "figures")
}
}
