% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUnify-package.R
\docType{package}
\name{scUnify-package}
\alias{scUnify}
\alias{scUnify-package}
\title{scUnify: Unification of single-cell multi-modal processing and visualization tools.}
\description{
Unification of single-cell multi-modal processing and visualization tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hungms.github.io/scUnify/}
}

}
\author{
\strong{Maintainer}: Miu Shing Hung \email{matthewhungms@gmail.com} (\href{https://orcid.org/0009-0004-0412-4374}{ORCID})

}
\keyword{internal}
