% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diffexp.R
\name{select_significant_genes}
\alias{select_significant_genes}
\title{Select significant genes from a dataframe}
\usage{
select_significant_genes(df, thresh = 0.5, gene_column = "gene", n = NULL)
}
\arguments{
\item{df}{A dataframe with padj and log2FoldChange columns}

\item{thresh}{A numeric threshold for the padj column}

\item{gene_column}{A character string specifying the column name of the gene IDs}

\item{n}{An optional numeric value specifying the number of genes to return}
}
\value{
A vector of gene IDs
}
\description{
This function selects significant genes from a dataframe based on the padj and log2FoldChange columns.
}
\examples{
\dontrun{
select_significant_genes(df, thresh = 0.5, gene_column = "gene", n = NULL)
}

}
