% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_diffexp.R
\name{select_top_deg}
\alias{select_top_deg}
\title{select_top_deg}
\usage{
select_top_deg(
  markers,
  n = 5,
  rank_by = "avg_log2FC",
  only.pos = F,
  min.pct = 0.1
)
}
\arguments{
\item{markers}{a dataframe of Seurat::FindAllMarkers output}

\item{n}{no. of genes per cluster}

\item{rank_by}{variables to rank by, either "avg_log2FC", "p_val_adj", "diff_pct"}

\item{only.pos}{if TRUE, return only genes with positive avg_log2FC. Defaults to FALSE}
}
\description{
Find top differentially expressed genes for each cluster from Seurat::FindAllMarkers output.
}
