% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_assays.R
\name{set_assay_keys}
\alias{set_assay_keys}
\title{set_assay_keys}
\usage{
set_assay_keys(obj)
}
\arguments{
\item{obj}{A Seurat object}
}
\value{
The Seurat object with standardized assay keys
}
\description{
Sets standardized keys for all assays in a Seurat object.
This function addresses inconsistencies in assay key naming by
setting each assay key to the lowercase assay name followed by an underscore.
}
\keyword{internal}
