% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_vdj.R
\name{seurat_add_dandelion}
\alias{seurat_add_dandelion}
\title{seurat_add_dandelion}
\usage{
seurat_add_dandelion(x, vdj, paired = T)
}
\arguments{
\item{x}{Seurat object}

\item{vdj}{dataframe of "all/filtered_contig_dandelion.tsv"}

\item{paired}{if TRUE, keep cells with a single heavy and a single light chain sequence}
}
\description{
add dandelion output to seurat object metadata
}
