% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_objects.R
\name{split_object}
\alias{split_object}
\title{split_object}
\usage{
split_object(obj, split.by)
}
\arguments{
\item{obj}{A Seurat object.}

\item{split.by}{A metadata column name to split the object by.}
}
\value{
A list of Seurat objects.
}
\description{
Splits a Seurat object into a list of Seurat objects based on a metadata column.
}
