% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_features.R
\name{subset_features}
\alias{subset_features}
\title{subset_features}
\usage{
subset_features(obj, features = NULL, assay.from, assay.to)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{a vector features to remove}

\item{assay.from}{current assay name}

\item{assay.to}{new assay name to store removed features}
}
\description{
Subset features from current assay to new assay
}
