% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_features.R
\name{subset_vdj_features}
\alias{subset_vdj_features}
\title{subset_vdj_features}
\usage{
subset_vdj_features(obj, assay, bcr = T, tcr = T)
}
\arguments{
\item{obj}{Seurat object}

\item{bcr}{if TRUE, remove BCR-VDJ features. Defaults to FALSE}

\item{tcr}{if TRUE, remove TCR-VDJ features. Defaults to FALSE}

\item{from.assay}{current assay name, defaults to "RNA"}
}
\description{
subset VDJ features from an assay
}
