% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_validations.R
\name{validate_assays}
\alias{validate_assays}
\title{Validate assay}
\usage{
validate_assays(obj)
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{Assay to validate}
}
\value{
Assay
}
\description{
This function validates that the assay is present in the Seurat object.
}
