% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_validating.R
\name{validate_object}
\alias{validate_object}
\title{Validate object}
\usage{
validate_object(obj)
}
\arguments{
\item{obj}{Seurat object}
}
\value{
Seurat object
}
\description{
This function validates that the object is a Seurat object.
}
