% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_validating.R
\name{validate_split.by}
\alias{validate_split.by}
\title{Validate split.by}
\usage{
validate_split.by(split.by, obj)
}
\arguments{
\item{split.by}{Split.by to validate}
}
\value{
Split.by
}
\description{
This function validates that the split.by is a character vector.
}
