% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_storage.R
\name{write_matrix}
\alias{write_matrix}
\title{Write a matrix to a directory}
\usage{
write_matrix(matrix, dir)
}
\arguments{
\item{matrix}{The matrix to write}

\item{dir}{The directory to write the matrix to}
}
\value{
None. Creates a matrix.mtx, genes.tsv, and barcodes.tsv file in the specified directory
}
\description{
This function writes a matrix to a directory, including the matrix, genes, and barcodes.
}
