% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_storage.R
\name{write_png}
\alias{write_png}
\title{Write a png file}
\usage{
write_png(plot, output_dir, filename, ...)
}
\arguments{
\item{plot}{The plot to write}

\item{output_dir}{The directory to write the plot to}

\item{filename}{The name of the plot}

\item{...}{Additional arguments to pass to the png function}
}
\value{
None. Creates a png file in the specified directory
}
\description{
This function writes a png file to a directory.
}
