% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b-CELLCYCLE_util.R
\name{cc_by_tricycle}
\alias{cc_by_tricycle}
\title{Calculate cell cycle phase with tricycle}
\usage{
cc_by_tricycle(obj, assay = "RNA", output_dir = NULL)
}
\description{
Calculate cell cycle phase with tricycle
}
