% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_objects.R
\name{create_seurat_object}
\alias{create_seurat_object}
\title{create_seurat_object}
\usage{
create_seurat_object(
  data_dirs,
  samples,
  format = "cellranger",
  hto_pattern = NULL,
  denoise_adt = TRUE
)
}
\arguments{
\item{data_dirs}{A vector of directories containing cellranger outputs}

\item{samples}{A vector of sample names for each directory}

\item{format}{The format of the data to be loaded. Either "cellranger" or "h5". Defaults to "cellranger"}

\item{hto_pattern}{Prefix to identify HTO tag names if HTO library is present. Defaults to NULL}

\item{denoise_adt}{If TRUE, denoise ADT counts using DSB normalization. Defaults to TRUE}
}
\value{
A list of Seurat objects, with each entry corresponding to a sample
}
\description{
Creates a list of Seurat objects from 10x Genomics cellranger-multi outputs.
This function processes RNA, ADT (Antibody-Derived Tags), and HTO (Hashtag Oligonucleotides)
modalities if present in the cellranger outputs.
}
