% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_tools.R
\name{create_seurat_object}
\alias{create_seurat_object}
\title{create_seurat_object}
\usage{
create_seurat_object(dir, samples, hto_str = NULL, adt_normalize = T)
}
\arguments{
\item{dir}{a vector of directories}

\item{samples}{a vector of sample names for each directory}

\item{hto_str}{prefix to identify HTO tag names if HTO library is present. Defaults to NULL}
}
\description{
Create Seurat object from cellranger-multi outputs
}
