% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{dotplot_get_expression}
\alias{dotplot_get_expression}
\title{Get gene expression data}
\usage{
dotplot_get_expression(
  obj,
  genes,
  group.by,
  split.by = NULL,
  assay = "RNA",
  slot = "data",
  scale = TRUE
)
}
\arguments{
\item{obj}{Seurat object}

\item{genes}{Vector of genes to include}

\item{group.by}{Column to group by}

\item{split.by}{Column to split by (optional)}

\item{assay}{Assay to use}

\item{slot}{Data slot to use}

\item{scale}{Logical; whether to scale data}
}
\value{
Dataframe with summarized expression values
}
\description{
Extracts and summarizes gene expression data from a Seurat object
}
\keyword{internal}
