% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_features.R
\name{get_all_features}
\alias{get_all_features}
\title{Get all features from a Seurat object}
\usage{
get_all_features(obj)
}
\arguments{
\item{obj}{A Seurat object}
}
\value{
A vector of all features from the Seurat object
}
\description{
This function returns a vector of all features from a Seurat object, including
both gene names and cell names.
}
