% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_assays.R
\name{get_assay_keys}
\alias{get_assay_keys}
\title{get_assay_keys}
\usage{
get_assay_keys(obj)
}
\arguments{
\item{obj}{Seurat object}
}
\value{
A vector of assay keys
}
\description{
Get assay keys from a Seurat object
}
