% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduction.R
\name{gg_features}
\alias{gg_features}
\title{gg_features}
\usage{
gg_features(
  plot,
  df,
  group.by = NULL,
  split.by = NULL,
  pal = NULL,
  pt.size = 0.5,
  pt.alpha = 0.5,
  legend.ncol = 1,
  label = T,
  count = T,
  count.groups = T,
  mask = T,
  count.text.size = 5,
  label.size = 4,
  mask.size = 1
)
}
\arguments{
\item{plot}{ggplot object}

\item{df}{Data frame}

\item{pal}{Colors to use for features}

\item{pt.size}{Point size}

\item{pt.alpha}{Point alpha}

\item{legend.ncol}{Number of columns for legend}

\item{label}{Whether to label features}

\item{count}{Whether to count features}

\item{count.groups}{Whether to count features per group}

\item{mask}{Whether to mask features}

\item{count.text.size}{Text size for cell count}

\item{label.size}{Label size}

\item{mask.size}{Mask size}

\item{features}{Features to add}

\item{feature.size}{Feature size}
}
\value{
ggplot object
}
\description{
Add features to ggplot
}
\keyword{internal}
