% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduction.R
\name{gg_reduction}
\alias{gg_reduction}
\title{gg_reduction}
\usage{
gg_reduction(
  df,
  group.by = NULL,
  split.by = NULL,
  reduction.type = "UMAP",
  pt.size = 0.5,
  pt.alpha = 0.5,
  facet.ncol = 2,
  legend.ncol = 1,
  shuffle = T
)
}
\arguments{
\item{obj}{Seurat object}

\item{reduction}{Reduction name}
}
\value{
Data frame with reduction
}
\description{
Get reduction from Seurat object
}
\keyword{internal}
