% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_assays.R
\name{join_layers}
\alias{join_layers}
\title{join_layers}
\usage{
join_layers(obj)
}
\arguments{
\item{obj}{A Seurat object containing one or more Assay5 objects}
}
\value{
The Seurat object with all Assay5 layers joined
}
\description{
Joins all layers for each Assay5 object in a Seurat object.
This function identifies all assays of class Assay5 and applies JoinLayers to each one.
}
