% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_tools.R
\name{join_layers}
\alias{join_layers}
\title{join_layers}
\usage{
join_layers(x, assays = c("RNA", "BCR", "TCR", "CC", "ADT", "HTO"))
}
\arguments{
\item{x}{Seurat object}

\item{assays}{A vector of assay names to join. Defaults to {assays = c("RNA", "BCR", "TCR", "CC", "ADT", "HTO")}}
}
\value{
Seurat object
}
\description{
Join layers for multiple assays
}
