% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plotting.R
\name{plot_dotplot}
\alias{plot_dotplot}
\title{plot_dotplot}
\usage{
plot_dotplot(
  x,
  group.by,
  split.by = NULL,
  features,
  assay = "RNA",
  slot = "data",
  scale = T,
  palette = "RdBu",
  direction = -1,
  diffexp = NULL,
  coord_flip = T,
  unique = F
)
}
\arguments{
\item{x}{Seurat object}

\item{group.by}{metadata column to group by}

\item{split.by}{metadata column to group by}

\item{features}{feature names stored in a vector or in a named list}

\item{assay}{assay name, defaults to "RNA"}

\item{slot}{slot name, defaults to "data"}

\item{scale}{if TRUE, scale gene expression across cells. Defaults to TRUE}

\item{palette}{a vector of colors, defaults to "Reds" from RColorBrewer}

\item{direction}{direction of palette, default to 1}

\item{diffexp}{whether to plot significance, input can be either a dataframe or a csv file path of the Seurat::FindAllMarkers output. Defaults to NULL.}

\item{coord_flip}{If TRUE, plot features on y-axis. Defaults to TRUE}

\item{unique}{If TRUE, do not repeat feature names. Defaults to FALSE}
}
\description{
Plot DotPlot from Seurat object
}
