% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plotting.R
\name{plot_percent}
\alias{plot_percent}
\title{plot_percent}
\usage{
plot_percent(
  x,
  group.by,
  variable,
  facet.by = NULL,
  cols = NULL,
  legend.ncol = 1
)
}
\arguments{
\item{x}{Seurat object}

\item{group.by}{column to group by}

\item{variable}{variable to color by}

\item{facet.by}{column to facet by}

\item{cols}{vector of colors}

\item{legend.ncol}{no. of legend columns}
}
\description{
plot percentage for cells
}
