% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduction.R
\name{plot_reduction}
\alias{plot_reduction}
\title{plot_reduction}
\usage{
plot_reduction(
  obj,
  group.by,
  reduction = "umap",
  reduction.type = "UMAP",
  split.by = NULL,
  pal = NULL,
  pt.size = 0.5,
  pt.alpha = 0.5,
  count = T,
  count.text.size = 5,
  count.groups = F,
  label = T,
  label.size = 4,
  facet.ncol = 2,
  legend.ncol = 1,
  shuffle = T
)
}
\arguments{
\item{obj}{Seurat object}

\item{group.by}{Metadata column to group by}

\item{reduction}{Reduction name}

\item{reduction.type}{Reduction type}

\item{split.by}{Metadata column to split by}

\item{pal}{Colors to use for groups}

\item{pt.size}{Point size}

\item{pt.alpha}{Point alpha}

\item{count}{Whether to count cells}

\item{count.text.size}{Text size for cell count}

\item{count.groups}{Whether to count cells per group}

\item{label}{Whether to label groups}

\item{label.size}{Label size}

\item{facet.ncol}{Number of columns for facets}

\item{legend.ncol}{Number of columns for legend}

\item{shuffle}{Whether to shuffle points}
}
\value{
ggplot object
}
\description{
Plot reduction from Seurat object
}
\examples{
\dontrun{
plot_reduction(obj, group.by = "seurat_clusters")
}
}
