% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting_reduction.R
\name{plot_reduction_density}
\alias{plot_reduction_density}
\title{plot_reduction_density}
\usage{
plot_reduction_density(
  obj,
  split.by = NULL,
  facet.ncol = 2,
  reduction = "umap",
  reduction.type = "UMAP",
  adjust = 5,
  pal = "Greys",
  ...
)
}
\arguments{
\item{obj}{Seurat object}

\item{split.by}{Metadata column to split by}

\item{facet.ncol}{Number of columns for facets}

\item{reduction}{Reduction name}

\item{reduction.type}{Reduction type}

\item{adjust}{Adjustment for density}

\item{pal}{Color palette}

\item{...}{Additional arguments to plot_reduction}
}
\value{
ggplot object
}
\description{
Plot reduction from Seurat object with density
}
\examples{
\dontrun{
plot_reduction_density(obj, split.by = "seurat_clusters")
}
}
