% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01-RNA_01-QC.R
\name{qc_by_mad}
\alias{qc_by_mad}
\title{qc by MAD}
\usage{
qc_by_mad(
  obj,
  split.by = NULL,
  var = c("nFeature_RNA", "nCount_RNA", "percent.mt", "percent.hb"),
  dev = 5,
  output_dir = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{var}{Variable columns from meta.data to calculate MAD for}

\item{dev}{Deviation from median}

\item{output_dir}{Directory to save MAD threshold table and outcomes}
}
\value{
Seurat object
}
\description{
This function calculates the MAD of features in the Seurat object.
}
