% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_features.R
\name{remove_features}
\alias{remove_features}
\title{remove_features}
\usage{
remove_features(obj, features = NULL, from.assay = "RNA", to.assay)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{a vector features to remove}

\item{from.assay}{current assay name, defaults to "RNA"}

\item{to.assay}{new assay name to store removed features}
}
\description{
Remove features from current assay to new assay
}
