% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_processing.R
\name{run_multiseqdemux}
\alias{run_multiseqdemux}
\title{run_multiseqdemux}
\usage{
run_multiseqdemux(x, process = F, add.one = F)
}
\arguments{
\item{x}{Seurat object}

\item{process}{if TRUE, run cluster algorithms on HTO counts}

\item{add.one}{+1 pseudocount for each hashtag oligo in each cell}
}
\description{
MULTIseqDemux algorithm to demultiplex cell hashtags
}
